Critical Information Update!
There was a missing piece in the algorithms (both recurrence and analytical versions). The missing piece was that it didn't make sense why I would include the "bank share" amount as part of the current capital. I revisited the code and the formulas and I found the error. I forgot to DOCUMENT the part that deducts the bank share from the net profit. The simulations are ALL CORRECT but the PDF didn't reflect that and could have caused a HUGE misunderstanding.